requirements.txt
Flask==2.1.2
python-dotenv==0.19.2
pymongo==4.7.2
Flask-PyMongo==2.3.0